OS X: Misc Fixes
This patch adds a series of smaller fixes to compile on OS X

  * Move to gutil/walltime.h for stopwatch
  * More size_t related modifications
  * Endieness is defined in a different header in OS X
  * sys/vfs.h does not exist, but the contents are provided by sys/mount.h
  * HOST_NAME_MAX is not defined on OS X

+++ be/src/util/collection-metrics.h
@@ -162,7 +162,7 @@ class StatsMetric : public Metric {
     container.AddMember("units", units, document->GetAllocator());
 
     if (StatsSelection & StatsType::COUNT) {
-      container.AddMember("count", boost::accumulators::count(acc_),
+      container.AddMember("count", (uint64_t) boost::accumulators::count(acc_),
           document->GetAllocator());
     }
 
@@ -198,7 +198,7 @@ class StatsMetric : public Metric {
     rapidjson::Value container(rapidjson::kObjectType);
 
     if (StatsSelection & StatsType::COUNT) {
-      container.AddMember("count", boost::accumulators::count(acc_),
+      container.AddMember("count", (uint64_t) boost::accumulators::count(acc_),
           document->GetAllocator());
     }
